// src/features/auth/auth.schema.ts
import { z } from "zod";

export const loginSchema = z.object({
	email: z.string().email("Email inválido"),
	password: z.string().min(8, "Mínimo 8 caracteres"),
});

export const signupSchema = z.object({
	name: z.string().min(1, "Nombre requerido"),
	email: z.string().email("Email inválido"),
	password: z.string().min(8, "Mínimo 8 caracteres"),
});

export type LoginInput = z.infer<typeof loginSchema>;
export type SignupInput = z.infer<typeof signupSchema>;